requirements.txt
flask
requests
gunicorn
python-dotenv
python-dateutil
flask-jwt-extended
flasgger
